+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2025-07-18
 
 ### Added
 
-- `GridRead` has default implementations of `{row|col|rect}_iter`.
+- Added (with default implementations) to `GridRead`, `GridReadUnchecked`
+  - `rect_iter`, `rect_iter_unchecked`
+- Added (with default implementations) to `GridWrite`, `GridWriteUnchecked`
+  - `fill_rect`, `fill_rect_unchecked`
+  - `fill_rect_iter`, `fill_rect_iter_unchecked`
+  - `fill_rect_solid`, `fill_rect_solid_unchecked`
 
 ### Removed